⚡ Replace synchronous HttpURLConnection with non-blocking OkHttp requests
Replaced blocking HttpURLConnection network calls with non-blocking OkHttp calls via suspendCancellableCoroutine in AnnouncementPreGenerator.kt, and deduplicated coroutine logic using OkHttpExtensions.kt.
